refactor(ptv3): move SemsegPreprocessCuda to its own file
Relocate SemsegPreprocessCuda (and SemsegSourceReconstructionView) from
preprocess_kernel.{hpp,cu} into preprocess/semseg_preprocess_kernel.{hpp,cu}.

Pure move: the deleted block in preprocess_kernel matches the new translation
unit byte-for-byte. The backbone preprocessor kernel keeps only its own code.
