+++ perception/autoware_ptv3/CMakeLists.txt
@@ -83,6 +83,7 @@ if(TRT_AVAIL AND CUDA_AVAIL)
     lib/postprocess/postprocess_kernel.cu
     lib/postprocess/detection3d_postprocess.cu
     lib/preprocess/preprocess_kernel.cu
+    lib/preprocess/semseg_preprocess_kernel.cu
   )
 
   target_include_directories(${PROJECT_NAME}_cuda_lib
