+++ perception/autoware_ptv3/include/autoware/ptv3/preprocess/preprocess_kernel.hpp
@@ -22,7 +22,6 @@
 
 #include <cuda_runtime_api.h>
 
-#include <cstddef>
 #include <cstdint>
 #include <vector>
 
@@ -104,41 +103,6 @@ private:
   autoware::cuda_utils::CudaUniquePtr<std::int64_t[]> pooling_run_ids_d_{nullptr};
   autoware::cuda_utils::CudaUniquePtr<std::uint8_t[]> pooling_workspace_d_{nullptr};
   std::size_t pooling_workspace_size_{0};
-};
-
-/**
- * Device buffers requested by the semantic-segmentation module during preprocessing.
- *
- * The backbone preprocessor always produces voxelized backbone inputs. When semseg output is
- * requested, this view points at semseg-owned buffers that preserve enough source-point information
- * to reconstruct voxel predictions back into the selected source point domain.
- */
-struct SemsegSourceReconstructionView
-{
-  float * reconstruction_features{};
-  void * cropped_source_points{};
-  std::int64_t * inverse_map{};
-  std::size_t * num_cropped_points{};
-};
-
-class SemsegPreprocessCuda
-{
-public:
-  explicit SemsegPreprocessCuda(const PTv3SemsegConfig & config);
-
-  /**
-   * Populate semseg-owned source reconstruction buffers from backbone preprocessing outputs.
-   *
-   * The caller passes the same execution stream used by the surrounding inference context. This
-   * helper does not store the stream because it is a leaf CUDA wrapper, not a pipeline phase owner.
-   */
-  void reconstructSource(
-    const PreprocessCuda & backbone_preprocessor, const void * input_data, CloudFormat input_format,
-    std::size_t num_points, std::size_t num_cropped_points, SemsegSourceReconstructionView view,
-    cudaStream_t stream) const;
-
-private:
-  PTv3SemsegConfig config_;
 };
 }  // namespace autoware::ptv3
